pyproject.toml
[tool.poetry]
name = "drf-proj"
version = "0.1.0"
description = ""
authors = ["Arthur Gaetov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"


[tool.poetry.group.dev.dependencies]
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
black = "^24.4.2"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
setuptools = "^70.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
